\name{Percentile.Plot}
\alias{Percentile.Plot}
\title{
Visualise percentile value}
\description{
This function will produce a plot with different coloured quartiles and a black dot representing the percentile value. }
\usage{
Percentile.Plot(percentile)
}
\arguments{
  \item{percentile}{
The desired percentile value}
}
\section{Error Messages}{
"Percentile value must be of class numeric"

The inputted percentile value must be numeric, without quotation marks.
}
