\name{Percentile_Calc}
\alias{Percentile_Calc}
\title{
Find percentile value }
\description{
This function will return the percentile of a value relative to a distribution. This distribution is created using values in a column of a given dataframe}
\usage{
Percentile_Calc(data, colnum, value)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
  Dataframe of interest.
  }
  \item{colnum}{
Column number of dataframe to create distribution with
}
  \item{value}{
The value you wish to get a percentile calculation for
}
}
