\name{Update.Package}
\alias{Update.Package}
\title{
Update the BoPRC package}
\description{
Updates the BoPRC package to the latest version. Suggestions for future updates should be sent to James Dare (James.Dare@boprc.govt.nz) or Data Services (EDS.JobRequests@boprc.govt.nz)}
\usage{
Update.Package()
}
